requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
faiss-cpu
langchain-groq
wikipedia
arxiv
langgraph
langgraph-cli[inmem]